switch js to ts (unchecked
