+++ backend/src/auth.ts
@@ -1,5 +1,5 @@
 import 'dotenv/config';
 
-export function authCheck(password) {
+export function authCheck(password: string): boolean {
   return password === process.env.AUTH_PASSWORD;
 }
