+++ backend/src/types/types.ts
@@ -1,0 +1,20 @@
+export interface DateParts {
+  dayString: string;
+  date: number;
+  month: number;
+  year: number;
+}
+
+export interface ReserveInfo {
+  username: string;
+  password: string;
+  date: DateParts;
+  startTimeIdx: number;
+  endTimeIdx: number;
+  courtOrder: string;
+}
+
+export interface PlaywrightResult {
+  success: boolean;
+  errorMessage: string;
+}
