--- backend/src/util.ts
@@ -1,3 +1,0 @@
-export function hello() {
-  console.log('hello wo sdfqrld');
-}
