+++ backend/src/utils/time.util.ts
@@ -1,7 +1,8 @@
-export function generateTimeOptions() {
-  const times = [];
-  let time = 8;
-  let hourString;
+export function generateTimeOptions(): string[] {
+  const times: string[] = [];
+  let time: number = 8;
+  let hourString: string;
+  
   while (time <= 22) {
     hourString = `${((time - 1) % 12) + 1 < 10 ? '0' : ''}${((time - 1) % 12) + 1}`;
     times.push(`${hourString}:00 ${time < 12 ? 'AM' : 'PM'}`);
