+++ frontend/src/types/reserve.type.ts
@@ -8,3 +8,5 @@ export interface ReserveInfo {
   endTimeIdx: number;
   courtOrder: string;
 }
+
+export interface ReserveResult
